+++ rust/crates/iqrah-core/src/exercises/pos_tagging_tests.rs
@@ -14,7 +14,7 @@ use std::collections::HashMap;
 struct MockContentRepo {
     texts: HashMap<i64, String>,                      // node_id -> text
     words: HashMap<String, Vec<Word>>,                // verse_key -> words
-    morphology: HashMap<i32, Vec<MorphologySegment>>, // word_id -> segments
+    morphology: HashMap<i64, Vec<MorphologySegment>>, // word_id -> segments
 }
 
 impl MockContentRepo {
@@ -283,7 +283,7 @@ impl ContentRepository for MockContentRepo {
         Ok(self.words.get(verse_key).cloned().unwrap_or_default())
     }
 
-    async fn get_word(&self, _word_id: i32) -> anyhow::Result<Option<Word>> {
+    async fn get_word(&self, _word_id: i64) -> anyhow::Result<Option<Word>> {
         Ok(None)
     }
 
@@ -326,7 +326,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_word_translation(
         &self,
-        _word_id: i32,
+        _word_id: i64,
         _translator_id: i32,
     ) -> anyhow::Result<Option<String>> {
         Ok(None)
@@ -410,7 +410,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_morphology_for_word(
         &self,
-        word_id: i32,
+        word_id: i64,
     ) -> anyhow::Result<Vec<MorphologySegment>> {
         Ok(self.morphology.get(&word_id).cloned().unwrap_or_default())
     }
@@ -463,8 +463,8 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_words_batch(
         &self,
-        word_ids: &[i32],
-    ) -> anyhow::Result<std::collections::HashMap<i32, crate::Word>> {
+        word_ids: &[i64],
+    ) -> anyhow::Result<std::collections::HashMap<i64, crate::Word>> {
         let mut result = std::collections::HashMap::new();
         for &id in word_ids {
             if let Some(word) = self.get_word(id).await? {
