+++ rust/crates/iqrah-core/src/exercises/reverse_cloze_tests.rs
@@ -265,7 +265,7 @@ impl ContentRepository for MockContentRepo {
         Ok(self.words.get(verse_key).cloned().unwrap_or_default())
     }
 
-    async fn get_word(&self, _word_id: i32) -> anyhow::Result<Option<Word>> {
+    async fn get_word(&self, _word_id: i64) -> anyhow::Result<Option<Word>> {
         Ok(None)
     }
 
@@ -308,7 +308,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_word_translation(
         &self,
-        _word_id: i32,
+        _word_id: i64,
         _translator_id: i32,
     ) -> anyhow::Result<Option<String>> {
         Ok(None)
@@ -392,7 +392,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_morphology_for_word(
         &self,
-        _word_id: i32,
+        _word_id: i64,
     ) -> anyhow::Result<Vec<MorphologySegment>> {
         Ok(vec![])
     }
@@ -445,8 +445,8 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_words_batch(
         &self,
-        word_ids: &[i32],
-    ) -> anyhow::Result<std::collections::HashMap<i32, crate::Word>> {
+        word_ids: &[i64],
+    ) -> anyhow::Result<std::collections::HashMap<i64, crate::Word>> {
         let mut result = std::collections::HashMap::new();
         for &id in word_ids {
             if let Some(word) = self.get_word(id).await? {
